[cli-dev-mode/optimizer] omit pageLoadAssetSizeLimit from cache (#95826)
