+++ packages/kbn-cli-dev-mode/src/cli_dev_mode.test.ts
@@ -135,6 +135,7 @@ it('passes correct args to sub-classes', () => {
           "repoRoot": <absolute path>,
           "runExamples": false,
           "silent": false,
+          "verbose": false,
           "watch": true,
         },
       ],
